scripts/requirements.txt
# Frame extraction and upload dependencies
opencv-python>=4.8.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
